test: split s3-spec test file into per-area suites
Splits packages/wgsl/tests/s3-spec-01-39.test.ts and
packages/wgsl/tests/s3-wgsl-runtime-more.test.ts into focused per-area
files: scanner, imports, cycles, resolver-paths, mangler, reflection,
validation, cache-key, export-attributes.

No production code or test logic changed; pure file reorganization.
Total test count preserved at 94.
