+++ packages/wgsl/tests/export-attributes.test.ts
@@ -1,0 +1,7 @@
+import { describe, expect, test } from "vitest";
+import { parseModule } from "../src/runtime/parser.ts";
+import { scan } from "../src/runtime/scanner.ts";
+
+describe("s3 §8 1-39", () => {
+  test("3 export attributes accepted and duplicate attribute rejected", () => { expect(parseModule(scan("@compute export fn main(){}" )).exports[0]?.name).toBe("main"); expect(parseModule(scan("export @compute fn main(){}" )).exports[0]?.name).toBe("main"); expect(() => parseModule(scan("@a export @b fn f(){}"))).toThrow(expect.objectContaining({ code: "VGPU-WGSL-EXP-NOTDECL" })); });
+});
